+++ src/main/java/fr/organizee/controller/ContactController.java
@@ -1,4 +1,4 @@
-package fr.organizee.dto.controller;
+package fr.organizee.controller;
 
 import fr.organizee.model.Contact;
 import fr.organizee.repository.ContactRepository;
