+++ src/main/java/fr/organizee/controller/TeamController.java
@@ -1,4 +1,4 @@
-package fr.organizee.dto.controller;
+package fr.organizee.controller;
 
 import fr.organizee.model.Team;
 import fr.organizee.repository.TeamRepository;
