+++ src/main/java/fr/organizee/controller/TodoListController.java
@@ -1,4 +1,4 @@
-package fr.organizee.dto.controller;
+package fr.organizee.controller;
 
 import fr.organizee.model.TodoList;
 import fr.organizee.repository.TodoListRepository;
